refactoring into leaner classes/methods/files
